Recreating Maya script Editor for Intellesense
Starting the basics of a script editor window to use for on the fly tool changes in Maya. the advantage is that it has intellesense for suggesting api commands while typing.  To me this just saves time with having to open copious chrome tabs to look up anything with any of the maya commands packages.